First commit
Reducers mostly working. Need to have reduce return a reducible, I
think. Something akin to the way fold returns a promise.
https://github.com/Gozala/reducers/blob/master/fold.js. What we're
dealing with is the repercussions of continuation passing style. You
need a way to resolve the continuation to a value, or else keep using
continuations.

I think the approach of returning values on down the chain and making
sure reduce can return a value (reducible) is the right approach.